[WR-398] default webauthn user verification to 'discouraged'
the webauthn spec defaults the user verification to 'preferred' which
causes some devices on some platforms to not behave correctly. in this
commit, i explicitly set the value to 'discouraged'.

if the server returns something other than 'not specified' it will instead
correctly set the webauthn user verification setting to that.